Custom types mapping

+++ client/arduino/SimpleSerialArduino/SimpleSerialSender.cpp
@@ -26,11 +26,11 @@ void SimpleSerialSender::sendInt(long value){
   Serial.write(dataOut,5);
 }
 void SimpleSerialSender::sendFloat(double value){
-  value *= 100;
+  value *= 100.;
   value += 32.768;
   dataOut[1] = id;
-  dataOut[2] = value / 256;
-  dataOut[3] = value % 256;
+  dataOut[2] = (long)value / 256;
+  dataOut[3] = (long)value % 256;
   Serial.write(dataOut,5);
 }
 void SimpleSerialSender::sendBool(bool value){
